requirements.txt
Pillow
numpy
tensorboardX
torch
torchvision
pytorch-ignite
